Rename banner component.
