+++ assets/js/modules/ads/components/notifications/index.js
@@ -19,3 +19,4 @@
 export { default as PAXSetupSuccessSubtleNotification } from './PAXSetupSuccessSubtleNotification';
 export { default as SetupSuccessSubtleNotification } from './SetupSuccessSubtleNotification';
 export { default as AccountLinkedViaGoogleForWooCommerceSubtleNotification } from './AccountLinkedViaGoogleForWooCommerceSubtleNotification';
+export { default as AdsModuleSetupCTABanner } from './AdsModuleSetupCTABanner';
